+++ Wancer/Wancer/Models/User.swift
@@ -10,16 +10,19 @@ import SwiftData
 
 @Model
 class User: Identifiable {
+    @Attribute(.unique)
     var id: Int
     var firstName: String
     var lastName: String
+    @Relationship(inverse: \Group.users)
+    var groups: [Group]
     
-    init(id: Int, firstName: String, lastName: String) {
+    init(id: Int, firstName: String, lastName: String, groups: [Group]) {
         self.id = id
         self.firstName = firstName
         self.lastName = lastName
+        self.groups = groups
     }
-    
 }
